+++ src/information/repositories/storeInformationRepository.js
@@ -10,7 +10,11 @@ module.exports = (informationItem) => new Promise((resolve, reject) => {
     mongoClient.connect('mongodb://localhost:27017/information-items')
         .then((database) => {
             const collection = database.db().collection('items');
-            collection.insertOne(informationItem)
+            collection.replaceOne(
+                { itemId: informationItem.itemId },
+                informationItem,
+                { upsert: true }
+            )
                 .then(response => {
                     resolve(response);
                 })