+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Added   | get route to query information-items
-Added   | post route to insert information-item
-Added   | put route to update information-item
+- Changed | query route now aggregates information from virtual items and information items
+- Added   | introduces virtual items
+- Added   | get route to query information-items
+- Added   | post route to insert information-item
+- Added   | put route to update information-item
 
 ## [Released]